Refactor: 폴더 구조 변경 및 네이밍컨벤션 SnakeCase로 변경-2
