+++ client/src/App.tsx
@@ -4,8 +4,8 @@ import { Route, Routes } from "react-router-dom";
 import Home from "@pages/home/home";
 import Cart from "@pages/cart/cart";
 import Place from "@pages/map/place";
-import Alcohol from "@pages/salesItems/Alcohol";
-import HelpCenter from "@pages/myPage/HelpCenter";
+import Alcohol from "@pages/salesItems/alcoholContainer";
+import HelpCenter from "@pages/myPage/helpCenter";
 import AlcoholDetail from "@pages/salesItems/AlcoholDetail";
 import Login from "@pages/login/login";
 import FindEmail from "@pages/login/findEmail";
@@ -13,16 +13,16 @@ import FindPassword from "@pages/login/findPassword";
 import SignupInput from "@pages/signUp/SignupInput";
 import SignupSelection from "@pages/signUp/SignupSelection";
 import SignupTerm from "@pages/signUp/SignupTerm";
-import LikePage from "@pages/myPage/LikeListPage";
-import OrderPage from "@pages/myPage/OrderListPage";
-import ChangeInfoPage from "@pages/myPage/ChangeInfoPage";
-import Payment from "@pages/payment/Payment";
-import PaymentConfirm from "@pages/payment/PaymentConfirm";
+import LikePage from "@pages/myPage/likeList";
+import OrderPage from "@pages/myPage/orderList";
+import ChangeInfoPage from "@pages/myPage/changeInfo";
+import Payment from "@pages/payment/payment";
+import PaymentConfirm from "@pages/payment/paymentConfirm";
 import FailPage from "@pages/payment/failPage";
 import CheckoutChang from "@pages/payment/chekOutModal";
 import ChatComponent from "@components/chat/chat";
 import Layout from "@layout/index";
-import ReviewEdit from "@pages/myPage/ReviewEdit";
+import ReviewEdit from "@pages/myPage/reviewEdit";
 
 const BodyContainer = styled.div`
   min-height: 100vh;
