+++ client/src/pages/Mypage/interface/index.ts
