+++ client/src/pages/Mypage/orderList.tsx
@@ -3,8 +3,8 @@ import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import { MdOutlineKeyboardArrowRight } from "react-icons/md";
 import { ButtonDark, ButtonLight } from "@components/common/commonButton";
-import Pagination from "@pages/salesItems/alcoholPage/Pagination";
-import * as Type from "./util";
+import Pagination from "@pages/salesItems/alcoholPage/pagination";
+import * as Type from "./interface";
 import * as styled from "./style";
 import { authTokenExpired } from "@utils/authExpired";
