--- client/src/pages/Mypage/util/index.ts
@@ -1,1 +1,0 @@
-export * from "./interface";
