+++ client/src/pages/Payment/Payment.tsx
@@ -4,9 +4,9 @@ import * as styled from "./style";
 import { ButtonDark, ButtonLight } from "@components/common/commonButton";
 import "react-datepicker/dist/react-datepicker.css";
 import Progress from "./Progress";
-import PaymnetUserInfo from "./PaymentUserinfo";
-import Itemlist from "./Paymentitemlist";
-import Payinfo from "./Paymentpayinfo";
+import PaymnetUserInfo from "./paymentUserInfo";
+import Itemlist from "./paymentIemList";
+import Payinfo from "./paymentPayInfo";
 import { UserProps } from "@utils/types";
 import { useSelector } from "react-redux";
 import axiosInstance from "@utils/api/axiosInstance";
