+++ client/src/pages/Signup/SignupInput.tsx
@@ -7,7 +7,7 @@ import { ButtonDark, ButtonLight } from "@components/common/commonButton";
 import Alert from "@components/common/commonAlert";
 import * as styled from "./styles";
 import * as Type from "@utils/types";
-import * as Common from "@styles/CommonConainer";
+import * as Common from "@styles/commonContainer";
 
 const url = `${process.env.REACT_APP_API_URL}`;
