+++ client/src/pages/Signup/SignupTerm.tsx
@@ -6,7 +6,7 @@ import { ButtonDark, ButtonLight } from "@components/common/commonButton";
 import Alert from "@components/common/commonAlert";
 import Term from "@components/signUpTerm/termComponent";
 import * as styled from "./styles";
-import * as Common from "@styles/CommonConainer";
+import * as Common from "@styles/commonContainer";
 
 const SignupTerm = () => {
   const navigate = useNavigate();
