+++ client/src/pages/salesItems/styles/index.ts
@@ -1,7 +1,6 @@
-export * from "./AlcoholStyle";
-export * from "./DetailStyle";
-export * from "./QuantityStyle";
-export * from "./ReviewListStlye";
-export * from "./PaginationStyle";
-export * from "./SortingStlye";
-export * from "./alcoholBuyStyle";
+export * from "./alcoholStyle";
+export * from "./detailStyle";
+export * from "./quantityStyle";
+export * from "./reviewListStyle";
+export * from "./paginationStyle";
+export * from "./sortingStyle";
