--- client/src/pages/salesItems/util/index.ts
@@ -1,1 +1,0 @@
-export * from "@pages/salesItems/util/interface";
