+++ client/src/styles/commonContainer.ts
