+++ client/src/styles/index.ts
@@ -1,2 +1,2 @@
 export * from "./Common";
-export * from "./CommonConainer";
+export * from "./commonContainer";
